% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Mapping observations in specific states in a specific year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{integer}{state.num}

\item{integer}{year}
}
\value{
a map of state with observations for year
}
\description{
Mapping observations in specific states in a specific year
}

