% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Create data frame as table and stores it in fars_read}
\source{
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System
}
\usage{
fars_read(filename)
}
\arguments{
\item{string}{a filename  + additional info about path. Check working directory}
}
\value{
a data frame tbl
}
\description{
Create data frame as table and stores it in fars_read
}

