% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads in data from different years and stores in dat which contains month and year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector}
}
\value{
a data frame tbl which includes month and year across different years
}
\description{
Reads in data from different years and stores in dat which contains month and year
}

