% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizing data to see how many observations per year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{vactor}{takes years as parameter}
}
\value{
a data frame tbl which shows number of observations grouped by year and month
}
\description{
Summarizing data to see how many observations per year and month
}

