% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Make filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer}
}
\value{
a filename with the specified year in the format "accident_%d.csv.bz2"
}
\description{
Make filename
}

